front: ui-charts: move generic utilities in common
