+++ front/ui/ui-charts/src/manchette/index.ts
@@ -20,4 +20,4 @@ export {
 export type { Waypoint, ProjectPathTrainResult, InteractiveWaypoint, ChartPath } from './types';
 
 export { positionMmToKm, positionKmToMm } from './utils';
-export { timeScaleToZoomValue, isInteractiveWaypoint } from './utils/helpers';
+export { isInteractiveWaypoint } from './utils/helpers';
