+++ front/ui/ui-charts/src/manchette/utils/helpers.ts
@@ -1,25 +1,12 @@
 import { type ReactNode } from 'react';
 
-import { clamp } from 'lodash';
-
 import {
   BASE_WAYPOINT_HEIGHT,
-  MAX_ZOOM_MS_PER_PX,
-  MAX_ZOOM_X,
-  MIN_ZOOM_MS_PER_PX,
-  MIN_ZOOM_X,
   MAX_ZOOM_Y,
   MIN_ZOOM_Y,
   MAX_ZOOM_MANCHETTE_HEIGHT_MILLIMETER,
 } from '../consts';
 import type { InteractiveWaypoint, Waypoint } from '../types';
-
-export const zoomValueToTimeScale = (slider: number) =>
-  MIN_ZOOM_MS_PER_PX * Math.pow(MAX_ZOOM_MS_PER_PX / MIN_ZOOM_MS_PER_PX, slider / 100);
-
-export const timeScaleToZoomValue = (timeScale: number) =>
-  (100 * Math.log(timeScale / MIN_ZOOM_MS_PER_PX)) /
-  Math.log(MAX_ZOOM_MS_PER_PX / MIN_ZOOM_MS_PER_PX);
 
 /**
  * min zoom is computed with manchette px height between first and last waypoint.
@@ -53,23 +40,6 @@ export const spaceScaleToZoomValue = (
   ((MAX_ZOOM_Y - MIN_ZOOM_Y) * Math.log(spaceScale / minZoomMillimeterPerPx)) /
     Math.log(maxZoomMillimeterPerPx / minZoomMillimeterPerPx) +
   MIN_ZOOM_Y;
-
-/** Zoom on X axis and center on the mouse position */
-export const zoomX = (
-  currentZoom: number,
-  currentOffset: number,
-  newZoom: number,
-  position: number
-) => {
-  const boundedZoom = clamp(newZoom, MIN_ZOOM_X, MAX_ZOOM_X);
-  const oldTimeScale = zoomValueToTimeScale(currentZoom);
-  const newTimeScale = zoomValueToTimeScale(boundedZoom);
-  const newOffset = position - ((position - currentOffset) * oldTimeScale) / newTimeScale;
-  return {
-    xZoom: boundedZoom,
-    xOffset: newOffset,
-  };
-};
 
 export const selectWaypointsToDisplay = (
   waypoints: Waypoint[],
