+++ front/ui/ui-charts/src/spaceTimeChart/utils/canvas.ts
@@ -253,35 +253,6 @@ export function drawAliasedQuadrilateral(
   }
 }
 
-export function drawAliasedRect(
-  imageData: ImageData,
-  { x, y }: Point,
-  width: number,
-  height: number,
-  [r, g, b]: RGBColor | RGBAColor,
-  scalingRatio = 1
-) {
-  x = Math.round(x * scalingRatio);
-  y = Math.round(y * scalingRatio);
-  width = Math.round(width * scalingRatio);
-  height = Math.round(height * scalingRatio);
-
-  const xMin = clamp(x, 0, imageData.width);
-  const yMin = clamp(y, 0, imageData.height);
-  const xMax = clamp(x + width, 0, imageData.width);
-  const yMax = clamp(y + height, 0, imageData.height);
-
-  for (let i = xMin; i < xMax; i++) {
-    for (let j = yMin; j < yMax; j++) {
-      const index = (j * imageData.width + i) * 4;
-      imageData.data[index] = r;
-      imageData.data[index + 1] = g;
-      imageData.data[index + 2] = b;
-      imageData.data[index + 3] = 255;
-    }
-  }
-}
-
 export type CanvasRect = {
   timeStart: Date;
   timeEnd: Date;
